+++ lily/include/paper-book.hh
@@ -67,11 +67,11 @@ public:
 
 protected:
   void classic_output_aux (SCM output,
-                           int *first_performance_number);
-  int output_aux (SCM output_channel,
-                  bool is_last,
-                  int *first_page_number,
-                  int *first_performance_number);
+                           long *first_performance_number);
+  long output_aux (SCM output_channel,
+                   bool is_last,
+                   long *first_page_number,
+                   long *first_performance_number);
 };
 
 DECLARE_UNSMOB (Paper_book, paper_book)
